# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion Comptable",
        "description": "API pour la gestion comptable avec journal des écritures et génération de balance comptable",
        "contact": {
            "name": "Support API",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer la liste de toutes les transactions",
                "description": "Afficher une liste des transactions.",
                "operationId": "ceac716339cb2455648899fac415b98f",
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Créer une nouvelle transaction comptable",
                "description": "Enregistrer une nouvelle transaction.",
                "operationId": "d3a7a8d8ab4eeb1b4fb29af77c0e98f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "description",
                                    "compte_debit_id",
                                    "compte_credit_id",
                                    "montant"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Paiement facture client"
                                    },
                                    "compte_debit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "compte_credit_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur est survenue lors de la création de la transaction"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer les détails d'une transaction",
                "description": "Afficher les détails d'une transaction spécifique.",
                "operationId": "a64a50b3720f06a3a57ca6d76fae5b12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Utilisateur": {
                "title": "Utilisateur",
                "description": "Modèle d'utilisateur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@example.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+33612345678"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "123 Rue de Paris"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "comptable",
                            "utilisateur"
                        ],
                        "example": "comptable"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Compte": {
                "title": "Compte",
                "description": "Modèle de compte comptable",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Banque"
                    },
                    "code": {
                        "type": "string",
                        "example": "512"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "passif",
                            "produit",
                            "charge"
                        ],
                        "example": "actif"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EntreeJournal": {
                "title": "EntreeJournal",
                "description": "Modèle d'entrée dans le journal comptable",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "transaction_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "utilisateur_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "transaction": {
                        "$ref": "#/components/schemas/Transaction"
                    },
                    "utilisateur": {
                        "$ref": "#/components/schemas/Utilisateur"
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "title": "Transaction",
                "description": "Modèle de transaction comptable",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-15"
                    },
                    "description": {
                        "type": "string",
                        "example": "Paiement facture client"
                    },
                    "compte_debit_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "compte_credit_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 3
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "compteDebit": {
                        "$ref": "#/components/schemas/Compte"
                    },
                    "compteCredit": {
                        "$ref": "#/components/schemas/Compte"
                    },
                    "entreeJournal": {
                        "$ref": "#/components/schemas/EntreeJournal"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}